Added Vaccines.csv and Vitamins.csv Added Vaccine and Vitamin Tab to ClinicView Methods to Add, Remove, and Update stock of Vaccines
WIP on recording Immunizations in adding Visit

+++ Archive/Iteration3/Vitamin.java
@@ -6,7 +6,7 @@
  * @version April 17th, 2025
  */
 public class Vitamin extends Dosable {
-    int dosageMG;
+    double dosageMG;
     int vitaminNeeded;
     int deficient;
 
@@ -28,7 +28,7 @@ public class Vitamin extends Dosable {
      *
      * @return Dosage in milligrams.
      */
-    public int getDosageMG() {
+    public double getDosageMG() {
         return dosageMG;
     }
 }
